Change V-73217 to require manual review
The STIG author's intent is to confirm whether the administrative accounts are not the same as the individual's everyday non-privileged account. The individual's everyday non-privileged account may not necessarily exist on the server or under the users group. Hence, it is difficult to automate concisely. Currently the describe block only confirms whether the accounts/groups in the local administrators group match the supplied inputs array. It does not address the STIG author's requirements. Hence, the describe block should be converted to a skip, requiring manual review, similar to method used for V-73225.

  describe "A manual review is required to verify that each user with administrative privileges has a separate account for user duties and one for privileged duties." do
    skip "A manual review is required to verify that each user with administrative privileges has a separate account for user duties and one for privileged duties."
  end